+++ Front/src/App.css
@@ -11,3 +11,13 @@
     margin: 0;
     box-sizing: border-box;
 }
+
+html,
+body {
+    margin: 0;
+    padding: 0;
+    height: 100%;
+    box-sizing: border-box;
+    scrollbar-width: none;
+    scroll-behavior:smooth;
+}
