+++ Front/src/components/AccountPages/Register.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Register() {
+  return (
+    <div>Register</div>
+  )
+}
+
+export default Register
