+++ Front/src/components/ErrorPages/error.css
@@ -1,0 +1,3 @@
+.container-fluid {
+    height: 100vh
+}
