+++ Front/src/components/HomePage/Home.jsx
@@ -5,8 +5,9 @@ function Home() {
   return (
     <>
       <div className='fs-1 w-100 p-4 d-flex flex-row justify-content-evenly'>
-        <Link to="/budget">Biudžeto puslapis</Link>
-        {/* <Link to="/signup">Sign up</Link> */}
+        Home page (help! need design)
+        <Link to="/dashboard">Dashboard</Link>
+        <Link to="/signup">Sign up</Link>
       </div>
     </>
   )
