--- Front/src/components/Main/Charts/chartInfo.css
@@ -1,39 +1,0 @@
-
-.balancesummary div:first-child>span:first-child>svg,
-.balancesummary div:nth-child(2)>span:first-child>svg,
-.balancesummary div:last-child>span:first-child>svg {
-    width: 2em;
-    height: 2em;
-    border-radius: 50%;
-    font-size: 1.3em;
-}
-
-/* .balancesummary div:first-child,
-.balancesummary div:nth-child(2),
-.balancesummary div:last-child {
-    width: 50%;
-} */
-.balancesummary div:first-child,
-.balancesummary div:nth-child(2),
-.balancesummary div:last-child {
-    width: 33%;
-    transition: 500ms linear;
-}
-
-.balancesummary svg {
-    color: white;
-}
-
-.balancesummary>div:nth-child(1)>span:nth-child(1)>span:nth-child(2),
-.balancesummary>div:nth-child(2)>span:nth-child(1)>span:nth-child(2),
-.balancesummary>div:nth-child(3)>span:nth-child(1)>span:nth-child(2) {
-    font-size: 1.5em;
-    color: rgb(77, 77, 77);
-}
-
-.balancesummary > div:nth-child(1) > span:nth-child(1) > svg:nth-child(1),
-.balancesummary > div:nth-child(2) > span:nth-child(1) > svg:nth-child(1),
-.balancesummary > div:nth-child(3) > span:nth-child(1) > svg:nth-child(1) {
-    box-shadow: 0 0 15px rgba(110, 61, 61, 0.8);
-    transition: 500ms linear;
-}
