--- Front/src/components/Main/Veikla/activitiesmain.css
@@ -1,168 +1,0 @@
-table tbody td {
-    vertical-align: baseline;
-    width: 1000px;
-    padding: 0.1em;
-    position: relative;
-}
-
-table thead tr {
-    border-bottom: 1px solid lightgrey;
-}
-
-tbody.text-center>tr:nth-child(1)>td:nth-child(3) {
-    min-width: 18%;
-}
-
-.cardicons svg {
-    border-radius: 50%;
-}
-
-.editinputs td input,
-.editinputs td select {
-    text-align: center;
-    border: 1px solid lightgrey;
-    background-color: transparent;
-    border-radius: 15px;
-    outline: darkgrey;
-    -webkit-transition: 700ms all;
-    -o-transition: 700ms all;
-    padding: 0.1em;
-}
-
-.editinputs td>select {
-    text-align: center;
-    outline: darkgrey;
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
-}
-
-.editinputs>td:nth-child(4) {
-    max-width: 650px;
-}
-
-.editinputs td>input[type=date] {
-    -webkit-appearance: none;
-    font-size: 1em;
-    color: #000;
-}
-
-.editinputs td input:focus,
-.editinputs td select:focus {
-    background: lightgrey;
-}
-
-
-.buttons button:first-child,
-.editbuttons button:first-child {
-    background-color: #ed99a2;
-    color: #dc3545;
-    transition: 500ms linear;
-}
-
-.buttons button:first-child:hover,
-.editbuttons button:first-child:hover {
-    background-color: #dc3545;
-    color: #ffc107;
-    transition: 500ms linear;
-}
-
-.buttons button:last-child,
-.editbuttons button:last-child {
-    background-color: #85b5fe;
-    color: #0d6efd;
-    transition: 500ms linear;
-}
-
-.buttons button:last-child:hover,
-.editbuttons button:last-child:hover {
-    background-color: #0d6efd;
-    color: #ffc107;
-    transition: 500ms linear;
-}
-
-.text-yellow {
-    color: #f2d98f;
-}
-
-.editinputs>td {
-    position: relative;
-}
-
-.error1,
-.error2,
-.error3,
-.error4 {
-    position: absolute;
-    width: 12em;
-    height: 7em;
-    font-size: 1em;
-    color: rgb(98, 0, 0);
-    background-color: #ebebeb;
-    text-align: center;
-    display: block;
-    font-weight: 500;
-    -webkit-box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
-    box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
-    border-radius: 20px;
-    z-index: 1;
-}
-
-.error1 {
-    left: 0.5em;
-    bottom: -8em;
-}
-
-.error3 {
-    left: 0.8em;
-    bottom: -8em;
-}
-
-.error4 {
-    left: 0.2em;
-    bottom: -8em;
-}
-
-.error1::before,
-.error2::before,
-.error3::before,
-.error4::before {
-    content: "";
-    position: absolute;
-    bottom: 100%;
-    left: 50%;
-    margin-left: -12px;
-    border-width: 12px;
-    border-radius: 2px;
-    border-style: solid;
-    border-color: transparent transparent #c6c6c6 transparent;
-}
-
-.navigation {
-    width: 1px;
-    border: 1px solid red;
-}
-
-div.justify-content-end:nth-child(1) > div:nth-child(2) {
-    margin: 0 2em;
-}
-
-.main-exp>div:nth-child(1)>span:before {
-    content: '';
-    display: inline-flex;
-    width: 2.2em;
-    height: 0.7em;
-    border-radius: 15px;
-    margin: 0 0.2em;
-    background-color: #0d6efd;
-}
-
-.main-exp>div:nth-child(2)>span:before {
-    content: '';
-    display: inline-flex;
-    width: 2.2em;
-    height: 0.7em;
-    border-radius: 15px;
-    margin: 0 0.2em;
-    background-color: #dc3545;
-}
