--- Front/src/components/Main/budgetmain.css
@@ -1,313 +1,0 @@
-html,
-body {
-    margin: 0;
-    display: block;
-    height: 100%;
-}
-
-.maincontent {
-    flex-basis: 85%;
-    flex-grow: 1;
-    min-height: 100vh;
-    background-color: #e8e8e8dc;
-}
-
-.sidemenu {
-    min-height: 100vh;
-    flex-basis: 15%;
-    z-index: 1000;
-    background-color: whitesmoke;
-    box-shadow: 3px 3px 15px #aeaeae;
-    padding-left: 0;
-    padding-right: 0;
-}
-
-.sidemenu a:nth-child(2) span,
-.sidemenu a:last-child span {
-    border-left: 5px solid #ffc107;
-    transition: 500ms linear;
-}
-
-.sidemenu a:nth-child(2),
-.sidemenu a:last-child {
-    transition: 500ms all;
-    width: 90%;
-}
-
-.sidemenu a:first-child {
-    color: rgb(112, 103, 50);
-}
-
-
-.sidemenu a:nth-child(2):hover,
-.sidemenu a:last-child:hover {
-    border-radius: 15px;
-    width: 105%;
-    background-color: #cfcfcf;
-    padding: 0;
-    transform: scale(1.1, 1.1);
-}
-
-.header {
-    background-color: whitesmoke;
-}
-
-.title {
-    font-weight: 400;
-    padding-left: 4em;
-    position: relative;
-    letter-spacing: 0.1em;
-}
-
-.title:after {
-    content: "";
-    border-bottom: 2px solid;
-    bottom: -6px;
-    left: 4em;
-    position: absolute;
-    width: 30px;
-}
-
-.account {
-    cursor: pointer;
-    position: relative;
-    flex-grow: 1;
-}
-
-.account span svg {
-    transition: 500ms all;
-}
-
-.acc-content {
-    background-color: whitesmoke;
-    position: absolute;
-    top: 3.5em;
-    right: 2.5em;
-    min-width: 7em;
-    min-height: 5em;
-    padding: 0.5em;
-}
-
-.netbalance {
-    width: 180px;
-    height: 180px;
-    padding: 2em;
-    border-radius: 50%;
-    text-align: center;
-    border: 6px solid grey;
-    font-size: 1em;
-}
-
-.main {
-    padding: 0 5em 0 5em;
-}
-
-.main .button div:first-child,
-.main .button div:nth-child(2),
-.main .button div:last-child {
-    width: 33%;
-}
-
-.main .button button {
-    border: none;
-    width: 2em;
-    height: 2em;
-    border-radius: 50%;
-    font-size: 1.5em;
-    background-color: lightslategrey;
-    transition: 500ms linear;
-}
-
-.button>div svg {
-    color: #fff;
-}
-
-.button>div:nth-child(1)>span:nth-child(2),
-.button>div:nth-child(2)>span:nth-child(2),
-.button>div:nth-child(3)>span:nth-child(2) {
-    font-size: 1.3em;
-    color: rgb(77, 77, 77);
-}
-
-.filter,
-.add,
-.download {
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.8);
-}
-
-.main .filter:hover {
-    background-color: rgb(66, 66, 66);
-    box-shadow: 0 0 30px rgb(66, 66, 66);
-}
-
-.main .add:hover {
-    background-color: rgb(50, 60, 112);
-    box-shadow: 0 0 30px rgb(50, 60, 112);
-}
-
-.main .download:hover {
-    background-color: rgb(112, 103, 50);
-    box-shadow: 0 0 30px rgb(112, 103, 50);
-}
-
-.main .activitiestable:first-of-type {
-    background-color: rgba(167, 167, 167, 0.192);
-}
-
-.activitiestable {
-    height: 100%;
-    border-radius: 20px;
-    font-size: 0.8em;
-}
-
-.popupform {
-    position: fixed;
-    justify-content: center;
-    align-items: center;
-    align-content: center;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100vh;
-    z-index: 1000;
-    background: #00000050;
-}
-
-.formblock {
-    background: #fff;
-    position: relative;
-    width: 38em;
-    height: auto;
-    border-radius: 10px;
-}
-
-
-.formblock .formtitle span:first-child,
-.formblock .formtitle span:last-child {
-    width: 8%;
-}
-
-.formblock .formtitle h4 {
-    width: 92%;
-}
-
-.formblock .formtitle span:last-child {
-    cursor: pointer;
-    font-size: 1.5em;
-}
-
-.formtitle div {
-    padding-left: 0.1em;
-    padding-right: 0.1em;
-    position: relative;
-}
-
-.formtitle div:first-child svg {
-    width: 90%;
-    height: auto;
-    font-weight: 900;
-    font-size: 1.3em;
-    padding-bottom: 0.1em;
-}
-
-
-.formtitle div:first-child:before {
-    content: '';
-    position: absolute;
-    bottom: -50%;
-    right: -62%;
-    border-radius: 50%;
-    width: 3em;
-    height: 3em;
-    background: #8080802f;
-}
-
-.outflowbtn,
-.inflowbtn {
-    border-radius: 50%;
-    border: none;
-    position: relative;
-    transition: 500ms linear;
-    background-color: lightgrey;
-}
-
-.outflowbtn svg,
-.inflowbtn svg {
-    color: #fff;
-    text-align: center;
-    font-size: 2em;
-    font-weight: 500;
-    height: auto;
-}
-
-.outflowbtn:hover {
-    background-color: #dc3545;
-}
-
-.inflowbtn:hover {
-    background-color: #0d6efd;
-
-}
-
-.formblock form label {
-    margin-left: 1.5em;
-    margin-bottom: 0.2em;
-}
-
-.formblock form input,
-.formblock form select {
-    width: 95%;
-    margin: 0 auto;
-    padding: 0.6em;
-    border-radius: 10px;
-    position: relative;
-    outline: none;
-    -webkit-appearance: none;
-    box-shadow: inset 0 0 5px 0 rgba(0, 0, 0, 0.295);
-    color: rgb(90, 90, 90);
-}
-
-.amountblock,
-.dateblock {
-    width: 47%;
-    margin: 0 auto;
-}
-
-input[type='date'] {
-    color: rgb(90, 90, 90);
-}
-
-.formblock select {
-    background-color: transparent;
-    -webkit-appearance: none;
-    box-shadow: inset 0 0 5px 0 rgba(0, 0, 0, 0.295);
-    outline: none;
-}
-
-.formfooter {
-    height: 8vh;
-    border-radius: 10px;
-    background-color: rgb(235, 235, 235);
-    align-items: center;
-    justify-content: end;
-}
-
-div.me-4:nth-child(1)>button:nth-child(1) {
-    background-color: #0d6efd;
-    transition: 500ms linear;
-    font-size: 1.1em;
-}
-
-div.me-4:nth-child(1)>button:nth-child(1):hover {
-    background-color: #0041a3;
-}
-
-div.me-4:nth-child(2)>button:nth-child(1) {
-    background-color: #d6d6d6d2;
-    transition: 500ms linear;
-    font-size: 1.1em;
-}
-
-div.me-4:nth-child(2)>button:nth-child(1):hover {
-    background-color: #afafafd2;
-}
