+++ Front/src/components/MainPages/Charts/chart.css
@@ -1,0 +1,78 @@
+.balancesummary div:first-child>span:first-child>svg,
+.balancesummary div:nth-child(2)>span:first-child>svg,
+.balancesummary div:last-child>span:first-child>svg {
+    width: 2em;
+    height: 2em;
+    border-radius: 50%;
+    font-size: 1.6em;
+}
+
+/* .balancesummary div:first-child,
+.balancesummary div:nth-child(2),
+.balancesummary div:last-child {
+    width: 50%;
+} */
+.balancesummary div:first-child,
+.balancesummary div:nth-child(2),
+.balancesummary div:last-child {
+    width: 33%;
+}
+
+.balancesummary svg {
+    color: white;
+}
+
+.balancesummary>div:nth-child(1)>span:nth-child(1)>span:nth-child(2),
+.balancesummary>div:nth-child(2)>span:nth-child(1)>span:nth-child(2),
+.balancesummary>div:nth-child(3)>span:nth-child(1)>span:nth-child(2) {
+    font-size: 1.2em;
+    color: rgb(77, 77, 77);
+}
+
+.balancesummary>div:nth-child(1)>span:nth-child(1)>svg:nth-child(1),
+.balancesummary>div:nth-child(2)>span:nth-child(1)>svg:nth-child(1),
+.balancesummary>div:nth-child(3)>span:nth-child(1)>svg:nth-child(1) {
+    -webkit-box-shadow: 0 0 15px rgba(110, 61, 61, 0.8);
+    box-shadow: 0 0 15px rgba(110, 61, 61, 0.8);
+-webkit-transition: 500ms linear;
+-o-transition: 500ms linear;
+transition: 500ms linear;
+}
+
+@media screen and (min-width:576px) and (max-width: 767px) {
+    
+    .balancesummary div {
+        font-size: 1em;
+    }
+
+}
+
+@media screen and (min-width:767px) and (max-width: 1370px) {
+
+    .balancesummary div:first-child,
+    .balancesummary div:nth-child(2),
+    .balancesummary div:last-child {
+        width: 18em;
+    }
+
+    .balancesummary div:first-child,
+    .balancesummary div:nth-child(2),
+    .balancesummary div:last-child {
+        margin-bottom: 0.8em;
+    }
+}
+
+@media screen and (min-width:1370px) and (max-width: 1717px) {
+
+    .balancesummary div:first-child,
+    .balancesummary div:nth-child(2),
+    .balancesummary div:last-child {
+        width: 20em;
+    }
+
+    .balancesummary div:first-child,
+    .balancesummary div:nth-child(2),
+    .balancesummary div:last-child {
+        margin-bottom: 0.8em;
+    }
+}
